moved lib/.ts files to src/.ts moved util.js to src/util.js and created a task "copy-js" to move it to lib removed spurious source-map installs (as we are using gulp-sourcemap instead) removed "use strict" calls, as TS generates strict JS anyway refactored .gitignore to relevant subdirectories moved types to typings so it aligns with TSD a bit better if we choose to use it later
