--- lib/.gitignore
@@ -1,1 +1,0 @@
-rules-parser.js
