+++ src/simulator.ts
@@ -15,9 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/// <reference path="../types/node.d.ts" />
-require('source-map-support').install();
-"use strict";
+/// <reference path="../typings/node.d.ts" />
 
 var Promise = require('promise');
 var uuid = require('node-uuid');
