+++ test/mail-test.ts
@@ -13,8 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/// <reference path="../types/node.d.ts" />
-require('source-map-support').install();
+/// <reference path="../typings/node.d.ts" />
 var bolt = (typeof(window) != 'undefined' && window.bolt) || require('../lib/bolt');
 var rulesSuite = bolt.rulesSuite;
 var secrets = require('./auth-secrets');
